Added CracenClient.hpp and CracenClientTest
Cracen Client ist the class, that provides the functionality to connect
with the cracenServer and enter a context/graph. It sits between the transpot
layer and the library layer and provides the functionality to send and receive
data from other cracenClients. To specify over which edges data has to be sent,
the concept of send_policies is introduced